+++ lua/codecompanion/adapters/copilot.lua
@@ -1,8 +1,8 @@
-local curl = require("plenary.curl")
-
 local config = require("codecompanion.config")
+local curl = require("plenary.curl")
 local log = require("codecompanion.utils.log")
 local openai = require("codecompanion.adapters.openai")
+local utils = require("codecompanion.utils.adapters")
 
 ---@alias CopilotOAuthToken string|nil
 local _oauth_token
@@ -60,7 +60,13 @@ local function get_token()
 
   for _, file_path in ipairs(file_paths) do
     if vim.uv.fs_stat(file_path) then
-      local userdata = vim.json.decode(vim.fn.readfile(file_path)[1])
+      local userdata = vim.fn.readfile(file_path)
+
+      if vim.islist(userdata) then
+        userdata = table.concat(userdata, " ")
+      end
+
+      local userdata = vim.json.decode(userdata)
       for key, value in pairs(userdata) do
         if string.find(key, "github.com") then
           return value.oauth_token
@@ -96,17 +102,6 @@ local function authorize_token()
 
   _github_token = vim.json.decode(request.body)
   return _github_token
-end
-
----Prepare data to be parsed as JSON
----@param data string | { body: string }
----@return string
-local prepare_data_for_json = function(data)
-  if type(data) == "table" then
-    return data.body
-  end
-  local find_json_start = string.find(data, "{") or 1
-  return string.sub(data, find_json_start)
 end
 
 ---@class Copilot.Adapter: CodeCompanion.Adapter
@@ -176,7 +171,7 @@ return {
     end,
     tokens = function(self, data)
       if data and data ~= "" then
-        local data_mod = prepare_data_for_json(data)
+        local data_mod = utils.clean_streamed_data(data)
         local ok, json = pcall(vim.json.decode, data_mod, { luanil = { object = true } })
 
         if ok then