+++ lua/codecompanion/types.lua
@@ -55,7 +55,7 @@
 ---@field augroup integer The autocmd group ID
 ---@field watch fun(self: CodeCompanion.Watchers, bufnr: number): nil Start watching a buffer
 ---@field unwatch fun(self: CodeCompanion.Watchers, bufnr: number): nil Stop watching a buffer
----@field get_changes fun(self: CodeCompanion.Watchers, bufnr: number): CodeCompanion.Change[]|nil Get changes since last check
+---@field get_changes fun(self: CodeCompanion.Watchers, bufnr: number): CodeCompanion.Change[]|nil Get the latest changes in the buffer
 
 ---@class CodeCompanion.WatcherState
 ---@field content string[] Complete buffer content
@@ -78,7 +78,7 @@
 ---@field intro_message? boolean Whether the welcome message has been shown
 ---@field messages? table The messages in the chat buffer
 ---@field parser vim.treesitter.LanguageTree The Tree-sitter parser for the chat buffer
----@field References CodeCompanion.Chat.References
+---@field references CodeCompanion.Chat.References
 ---@field refs? table<CodeCompanion.Chat.Ref> References which are sent to the LLM e.g. buffers, slash command output
 ---@field settings? table The settings that are used in the adapter of the chat buffer
 ---@field subscribers table The subscribers to the chat buffer