+++ queries/markdown/tools.scm
@@ -2,6 +2,12 @@
   (atx_heading) @role
 ) @content
 
+
+(section
+  (atx_heading) @role
+  (#eq? @role "Response")
+) @content
+
 (section
   (fenced_code_block
     (info_string) @lang